consolidate code into lambda_function
